Move common or similar rules to common config, choose based on NODE_ENV
Source map generation and CSS loaders
