+++ webpack.config.common.js
@@ -1,5 +1,48 @@
-// const env = process.env.NODE_ENV;
 const path = require("path");
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+const PostCssNormalize = require("postcss-normalize");
+const PostCssFlexbugsFixes = require("postcss-flexbugs-fixes");
+const PostCssPresetEnv = require("postcss-preset-env");
+
+const isProd = process.env.NODE_ENV === "production";
+const plugins = isProd
+  ? [
+      new MiniCssExtractPlugin({
+        filename: "[name]-[contentHash].css",
+        chunkFilename: "[name]-[chunkHash].css"
+      })
+    ]
+  : [];
+const postCssLoader = {
+  loader: "postcss-loader",
+  options: {
+    plugins: [
+      PostCssFlexbugsFixes(),
+      PostCssPresetEnv({
+        autoprefixer: {
+          flexbox: "no-2009"
+        },
+        stage: 3
+      }),
+      PostCssNormalize()
+    ]
+  }
+};
+
+const sassPostCssLoaders = isProd
+  ? (postCssLoader,
+    {
+      loader: "sass-loader",
+      options: {
+        sourceMap: true
+      }
+    })
+  : {
+      loader: "sass-loader",
+      options: {
+        sourceMap: true
+      }
+    };
 
 module.exports = {
   entry: {
@@ -19,12 +62,42 @@ module.exports = {
       utils: path.join(__dirname, "src/utils")
     }
   },
+  plugins,
+  devtool: "source-map",
   module: {
     rules: [
       {
         test: /\.(js|jsx)$/,
         exclude: /node_modules/,
         use: ["babel-loader"]
+      },
+      {
+        test: /\.module\.(s?css|sass)$/,
+        use: [
+          isProd ? MiniCssExtractPlugin.loader : "style-loader",
+          {
+            loader: "css-loader",
+            options: {
+              modules: true,
+              sourceMap: true
+            }
+          },
+          sassPostCssLoaders
+        ]
+      },
+      {
+        test: /\.(s?css|sass)$/,
+        exclude: /\.module\.(s?css|sass)$/,
+        use: [
+          isProd ? MiniCssExtractPlugin.loader : "style-loader",
+          {
+            loader: "css-loader",
+            options: {
+              sourceMap: true
+            }
+          },
+          sassPostCssLoaders
+        ]
       },
       {
         test: /\.(ttf|eot|woff|woff2|otf)$/,
